% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{calcdA_mat}
\alias{calcdA_mat}
\title{Partial cross-section area from width and height}
\usage{
calcdA_mat(w, h, zero = c("first", "minimum", "median"))
}
\arguments{
\item{w}{Matrix of widths}

\item{h}{Matrix of heights}

\item{zero}{Where to reference the zero value of dA?}
}
\description{
Calculate partial cross-section area from DAWG-formatted width and height matrices
}
