% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{calcdA_vec}
\alias{calcdA_vec}
\title{Calculate partial cross-section area from width and height vectors}
\usage{
calcdA_vec(w, h, zero = c("first", "minimum", "median"))
}
\arguments{
\item{w}{vector of widths}

\item{h}{vector of heights}

\item{zero}{Where to reference the zero value of dA?}
}
\description{
Calculate partial cross-section area from width and height vectors
}
