% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{ccf_lag}
\alias{ccf_lag}
\title{Returns a lagged swotlist, possibly with a new Ahat piece}
\usage{
ccf_lag(swotlist, Ahat = TRUE, verbose = FALSE)
}
\description{
Returns a lagged swotlist, possibly with a new Ahat piece
}
