% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{nc_list}
\alias{nc_list}
\title{Read a netcdf file to a list.}
\usage{
nc_list(file)
}
\arguments{
\item{file}{string providing the location of a netcdf file.}
}
\description{
Produces a list of all data contained in the netcdf, with their original names.
}
