% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{nc_reach}
\alias{nc_reach}
\title{Read reach-averaged observations and other relevant reach-level data from
 a netcdf file}
\usage{
nc_reach(file, good_only = FALSE)
}
\arguments{
\item{file}{string providing the location of a SWOT-like netcdf file.}

\item{good_only}{if TRUE, subset the data to only include data from reaches
specified as "good" a priori.}
}
\description{
Produces a list of swot-like matrices from reach-averaged slots of a netcdf,
  renamed for easier reference.
}
