% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{plot_DAWG}
\alias{plot_DAWG}
\title{Plot all time series of a DAWG-formatted matrix.}
\usage{
plot_DAWG(dawgmat)
}
\arguments{
\item{dawgmat}{a DAWG-formatted (space-down, time-across) matrix.}
}
\description{
Produces a ggplot from a space-down, time-across matrix.
}
