% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{realA0}
\alias{realA0}
\title{Actual A0 values across locations of a swotlist}
\usage{
realA0(swotlist, zero = c("first", "minimum", "median"))
}
\arguments{
\item{swotlist}{a list of swot-like matrices}

\item{zero}{Where to reference the zero value of dA?}
}
\description{
Returns a vector of unobservable (via SWOT) cross-sectional areas for a given swotlist
}
