% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{rezero_dA}
\alias{rezero_dA}
\title{Adjust the zero-reference of partial cross-sectional area data.}
\usage{
rezero_dA(dAmat, zero = c("first", "minimum", "median"))
}
\arguments{
\item{dAmat}{a DAWG-formatted matrix of partial cross-sectional area observations.}

\item{zero}{Where to reference the zero value of dA?}
}
\description{
Adjust the zero-reference of partial cross-sectional area data.
}
