% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sword_names.R
\name{sword_names}
\alias{sword_names}
\title{Get river names from a SWORD network}
\usage{
sword_names(sword, name_col = NULL, sort = TRUE, message = TRUE)
}
\arguments{
\item{sword}{A SWORD \code{sf} object, data frame, or tibble containing either
reaches or nodes.}

\item{name_col}{Optional name of the river-name column. If \code{NULL}, the
function attempts to identify a suitable name column automatically.}

\item{sort}{Logical. If \code{TRUE}, river names are returned alphabetically.
Default is \code{TRUE}.}

\item{message}{Logical. If \code{TRUE}, report the number of unique river names
and the percentage of the network with an available river name.
Default is \code{TRUE}.}
}
\value{
A character vector containing unique river names.
}
\description{
Returns the unique river names contained in a SWORD reach or node dataset.
The function automatically detects whether the input contains reaches or
nodes and reports the proportion of features for which a river name is
available.
}
\examples{
\dontrun{
sword <- sword_download(
  ...,
  feature = "Reach"
)

sword_names(sword)

names <- sword_names(
  sword,
  message = FALSE
)
}

}
\author{
Florian Betz
}
