% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sword_summary.R
\name{sword_summary}
\alias{sword_summary}
\title{Summarize a SWORD river network}
\usage{
sword_summary(
  sword,
  by = NULL,
  length_col = NULL,
  name_col = NULL,
  length_unit = c("km", "m")
)
}
\arguments{
\item{sword}{A SWORD \code{sf} object, data frame, or tibble containing either
reaches or nodes.}

\item{by}{Optional grouping variable. One of \code{NULL}, \code{"hack_order"}, or
\code{"river_name"}. If \code{NULL}, an overall network summary is returned.}

\item{length_col}{Optional name of the reach-length column. If \code{NULL},
the function attempts to identify a suitable length column automatically.
This argument is only used for reach data.}

\item{name_col}{Optional name of the river-name column. If \code{NULL},
the function attempts to identify a suitable name column automatically.}

\item{length_unit}{Unit used for reporting river length. Either \code{"km"}
or \code{"m"}. Default is \code{"km"}.}
}
\value{
A tibble containing summary statistics for the supplied SWORD object.
If \code{by} is specified, one row is returned for each group.
}
\description{
Provides a concise summary of a SWORD river network represented by either
reaches or nodes. The function automatically detects the feature type from
the presence of \code{reach_id} or \code{node_id}.

For reach data, the summary includes the number of reaches, total river
length, river-name coverage, and, if available, information on Hack order.
For node data, the summary includes the number of nodes, the number of
represented reaches, and river-name coverage.

The network can additionally be summarized by Hack order or river name.
}
\examples{
\dontrun{
reaches <- sword_download(
  ...,
  feature = "Reach"
)

sword_summary(reaches)

sword_summary(
  reaches,
  by = "river_name"
)

reaches <- sword_topology(reaches)

sword_summary(
  reaches,
  by = "hack_order"
)

nodes <- sword_download(
  ...,
  feature = "Node"
)

sword_summary(nodes)

sword_summary(
  nodes,
  by = "river_name"
)
}

}
