% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sword_topology.R
\name{sword_topology}
\alias{sword_topology}
\title{Computing network topology for the SWORD network data}
\usage{
sword_topology(network)
}
\arguments{
\item{network}{the river network from the SWORD database; this function requires v17c of the SWORD database}
}
\value{
returns an sf object with the
}
\description{
classifies the mainstem and gives unique identifiers to each tributary
}
\author{
Florian Betz
}
