% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{swot_A}
\alias{swot_A}
\title{Calculate area from A0 (a vector) and dA (a matrix).}
\usage{
swot_A(A0vec, dAmat, zero = "same")
}
\arguments{
\item{A0vec}{a vector of A0}

\item{dAmat}{a matrix of dA observations}

\item{zero}{If not "same" (the default), passed to rezero_dA before
calculating.}
}
\description{
Calculate area from A0 (a vector) and dA (a matrix).
}
