% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{swot_bamdata}
\alias{swot_bamdata}
\title{Create BAM-ready data from a swotlist.}
\usage{
swot_bamdata(swotlist, Qhat = NULL, ...)
}
\arguments{
\item{swotlist}{a list of SWOT observables}

\item{Qhat}{Prior guess of mean discharge. If NULL, will attempt to get
from \code{attr(swotlist, "QWBM")}}
}
\description{
Produces a bamdata object to be used with the bamr package.
}
