% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{swot_plot}
\alias{swot_plot}
\title{Plot all variables, or a subset thereof, as timeseries}
\usage{
swot_plot(swotlist, vars = "all")
}
\arguments{
\item{swotlist}{a list of swot-like matrices}

\item{vars}{Which variables to plot from within the swotlist.
Defaults to "all".}
}
\description{
Produces a ggplot of time series faceted by variable. Each facet is a DAWG
  plot (a la \code{plot_DAWG()})
}
