% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{swot_purge_nas}
\alias{swot_purge_nas}
\title{Purge all times or locations containing NAs}
\usage{
swot_purge_nas(swotlist, purge = c("times", "locs"))
}
\arguments{
\item{swotlist}{a list of swot-like matrices}

\item{purge}{Which of either "times" or "locs" to purge when removing NAs.}
}
\description{
Produces a swotlist with the all times (default) or locations that contain
 any NAs. This operation is performed for all matrices within the swotlist.
}
