% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swot_spacetime.R
\name{swot_spacetime}
\alias{swot_spacetime}
\title{Create a SWOT river space-time representation}
\usage{
swot_spacetime(
  swot_tibble,
  sword,
  variable = c("wse", "width", "area_total"),
  bin_days = 1,
  quality = c("usable", "good", "good_suspect", "all"),
  distance_col = NULL,
  distance_unit = c("km", "m"),
  aggregation = c("mean", "median"),
  plot = TRUE,
  tile_width = 1.5
)
}
\arguments{
\item{swot_tibble}{A SWOT RiverSP tibble containing node or reach
observations.}

\item{sword}{Corresponding SWORD nodes or reaches. May be an \code{sf} object,
data frame, or tibble.}

\item{variable}{Variable to display. One of \code{"wse"}, \code{"width"}, or
\code{"area_total"}.}

\item{bin_days}{Width of the temporal bins in days. Default is \code{1}.}

\item{quality}{Quality levels to retain. One of \code{"all"}, \code{"good"},
\code{"good_suspect"}, or \code{"usable"}. Default is \code{"usable"}.}

\item{distance_col}{Optional name of the SWORD hydrological-distance
column. If \code{NULL}, the function attempts to identify it automatically.}

\item{distance_unit}{Unit used for outlet distance. Either \code{"km"} or \code{"m"}.
Default is \code{"km"}.}

\item{aggregation}{Function used when multiple observations occur in the
same temporal bin at the same river location. Either \code{"mean"} or
\code{"median"}. Default is \code{"mean"}.}

\item{plot}{Logical. If \code{TRUE}, return a \code{ggplot} heatmap. If \code{FALSE},
return the aggregated space-time tibble.}

\item{tile_width}{plotting with of each data entry, default is 1.5}
}
\value{
A \code{ggplot} object if \code{plot = TRUE}, otherwise a tibble containing the
aggregated space-time observations.
}
\description{
Creates a longitudinal space-time representation of SWOT RiverSP
observations. SWOT observations are joined with the corresponding SWORD
network to obtain hydrological distance from the outlet. Observations are
aggregated into temporal bins and can be returned as a tibble or displayed
as a heatmap.
}
\examples{
\dontrun{
swot_spacetime(
  rhone_swot,
  rhone_sword,
  variable = "wse"
)

swot_spacetime(
  rhone_swot,
  rhone_sword,
  variable = "width",
  bin_days = 5
)
}

}
\author{
Florian Betz
}
