% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{swot_sset}
\alias{swot_sset}
\title{Subset dimensions of swot-like data.}
\usage{
swot_sset(swotlist, keeptimes = 0L, keeplocs = 0L)
}
\arguments{
\item{swotlist}{a list of swot-like matrices}

\item{keeptimes}{Indices of times (columns) to keep. Default is 0, which
keeps all times. Negative indices are allowed.}

\item{keeplocs}{Indices of locations (rows) to keep. Default is 0, which
keeps all locations. Negative indices are allowed.}
}
\description{
Returns a swotlist with all of the matrices of the input
 subset to keep only the specified times and location.
}
