% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{swot_tidy}
\alias{swot_tidy}
\title{Produce a tidy data.frame from a swotlist of matrices}
\usage{
swot_tidy(swotlist)
}
\arguments{
\item{swotlist}{A named list of matrices}
}
\description{
Produce a tidy data.frame from a swotlist of matrices
}
