% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{swot_timelag}
\alias{swot_timelag}
\title{Apply a timelag to the rows of matrices in a swotlist.}
\usage{
swot_timelag(swotlist, lags)
}
\arguments{
\item{swotlist}{a list of swot-like matrices}

\item{lags}{a vector with the same length as the number of locations (rows)
in the swotlist data, specifying the number of indices to shift each row}
}
\description{
Produces a swotlist with each matrices' rows shifted by an amount
  specified by \code{lag}.
}
