% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{swot_untidy}
\alias{swot_untidy}
\title{Produce a swotlist from a tidy data.frame}
\usage{
swot_untidy(swotdf)
}
\arguments{
\item{swotdf}{a list of swot-like matrices}
}
\description{
The inverse of swot_tidy. Produces a swotlist of matrices from a
  tidy data.frame of the sort produced by \code{swot_tidy}
}
