% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swotr.R
\name{swot_vec2mat}
\alias{swot_vec2mat}
\title{Convert a vector to a suitably-dimensioned matrix}
\usage{
swot_vec2mat(vec, pattern)
}
\arguments{
\item{vec}{a vector}

\item{pattern}{a matrix with the desired dimensions}
}
\description{
Produces a matrix of the same dimensions as \code{pattern}; excactly
 one of these dimensions must be equal to the length of \code{vec}.
 If \code{pattern} is square, an error is raised.
}
