% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aoi_rhone}
\alias{aoi_rhone}
\title{aoi rhone}
\format{
\subsection{sf}{
}
}
\source{
\url{https://www.hydrosheds.org/products/hydrobasins}
}
\usage{
aoi_rhone
}
\description{
This is a subset of the HydroBasins level 4 dataset for the Rhone basin (France, Switzerland). The original attributes of the HydroBasins
data are retained in the data even thouhg they are not used in the examples of the package. Coordinate system of the data is WGS84 (EPSG 4326).
}
\keyword{datasets}
