% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sword_download.R
\name{sword_download}
\alias{sword_download}
\title{Download SWORD river network data}
\usage{
sword_download(
  continent = "EU",
  network = NULL,
  category = "reaches",
  time_out = 1200,
  aoi = NULL,
  out_file,
  read = TRUE
)
}
\arguments{
\item{continent}{which continent to download, see also}

\item{network}{if known, you can filter directly by the network ID to e.g. obtain the network of a specific basin}

\item{category}{what to get from the SWORD data, reaches or nodes. Can be "reaches" or "nodes"}

\item{time_out}{increase timeout of 60 seconds to something more meaningful for GB scale data;
default is 20 minutes what should be sufficient even with slower internet connections.}

\item{aoi}{sf object with your area of interest as polygon}

\item{out_file}{output file of the selected reaches or nodes}

\item{read}{whether the downloaded data should be directly read as sf object or not; default is TRUE}
}
\value{
if read=TRUE, the function returns an sf object of the downloaded network, otherwise it just downloads the data and returns a
message with the location where the data is stored
}
\description{
Downloads SWORD river network data as foundation to work with the reaches and nodes of the SWOT river observations
}
\details{
Please note that the download links are hard coded, so consider looking in the source code if errors arise
}
\references{
\itemize{
\item Altenau et al. (2021): The Surface Water and Ocean Topography (SWOT) Mission River Database (SWORD):
A global river network for satellite data products". Water Resources Research. https://doi.org/10.1029/2021WR030054
}
}
\author{
Florian Betz
}
